// frontend_admin/src/app/shared/enums/configuracion.enum.ts
export enum ActionType {
  BUILD_URL = 0,
  INIT = 1,
  LOAD = 2,
  SELECT = 3,
  CHANGE = 4,
  SAVE = 5,
  RESET = 6,
  SEARCH = 7,
  NEW = 8,
  EXPORT = 9,
  SEND = 10,
  BUILD = 11,
  VALID = 12,
  END = 13,
  HIDE = 14,
  ACTION = 15,
  SUBMIT = 16,
  CLOSE = 17,
  EMITTER = 18,
  DELETE = 19,
  FILE_SEARCH = 20,
  FILE_LOAD = 21,
  FILE_DELETE = 22,
  FILE_SEND = 23,
  FILE_STATUS = 24,
  FORMAT = 25,
}